Move identity method from identity provider to client repo
Remove identity provider as it is not needed anymore. Update tests
in accordance with these changes.
