+++ clients/postgres/channels_test.go
@@ -5,7 +5,6 @@ import (
 	"testing"
 
 	"github.com/mainflux/mainflux/clients"
-	"github.com/mainflux/mainflux/clients/mocks"
 	"github.com/mainflux/mainflux/clients/postgres"
 	"github.com/stretchr/testify/assert"
 )
@@ -152,13 +151,11 @@ func TestChannelConnect(t *testing.T) {
 func TestChannelDisconnect(t *testing.T) {
 	email := "channel-disconnect@example.com"
 
-	idp := mocks.NewIdentityProvider()
-
 	clientRepo := postgres.NewClientRepository(db, testLog)
 	client := clients.Client{
 		ID:    clientRepo.ID(),
 		Owner: email,
-		Key:   idp.Key(),
+		Key:   clientRepo.Key(),
 	}
 	clientRepo.Save(client)
 
@@ -190,13 +187,11 @@ func TestChannelDisconnect(t *testing.T) {
 func TestChannelAccessCheck(t *testing.T) {
 	email := "channel-access-check@example.com"
 
-	idp := mocks.NewIdentityProvider()
-
 	clientRepo := postgres.NewClientRepository(db, testLog)
 	client := clients.Client{
 		ID:    clientRepo.ID(),
 		Owner: email,
-		Key:   idp.Key(),
+		Key:   clientRepo.Key(),
 	}
 	clientRepo.Save(client)
