+++ clients/service_test.go
@@ -25,9 +25,8 @@ func newService(tokens map[string]string) clients.Service {
 	clientsRepo := mocks.NewClientRepository()
 	channelsRepo := mocks.NewChannelRepository(clientsRepo)
 	hasher := mocks.NewHasher()
-	idp := mocks.NewIdentityProvider()
 
-	return clients.New(users, clientsRepo, channelsRepo, hasher, idp)
+	return clients.New(users, clientsRepo, channelsRepo, hasher)
 }
 
 func TestAddClient(t *testing.T) {
